After several months paused, the project was revisited...
- Refactored a bit of CSS
- Remembered why I stopped work on this project. The image files are a
total mess. There is no standard naming convention and no standard
width/height. The quality of the photographs range widely. Massaging
dschorr’s corpus of work to a workable set of images is a massive and
time consuming project in itself. Just to get the images to a point
where in I could begin automation with scripts is a serious manual
task.